+++ plugin/jetbrains/CHANGELOG.md
@@ -3,6 +3,10 @@
 # SourceMarker Changelog
 
 ## [Unreleased]
+### Added
+- Infinite scroll on traces portal ([#210](https://github.com/sourceplusplus/SourceMarker/issues/210))
+### Changed
+- Single page portal ([#254](https://github.com/sourceplusplus/SourceMarker/issues/254))
 ### Upgraded
 - Apache SkyWalking (8.2.0 -> 8.3.0)